requirements.txt
aiogram>=3.13,<4
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
pydantic-settings>=2.5,<3
redis>=5.0,<6
httpx>=0.27,<1
python-dateutil>=2.9,<3
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
